Complete Adding Rust as target
Completed adding Rust as a target. All of the fixtures have been tested
in a separate Rust project to verify that they build and successfully
run against the Har test endpoint.

All tests are running and passing, except for the snapshot that verifies
all available targets, not sure where to update that.