feat: heavily simplify multipart formdata snippet generation (#224)
## 🧰 Changes

In September of last year I reworked this library[^1] off `node-fetch`
and the funky non-native `FormData` object it ships to only support
native `FormData` -- in the process making code snippet generation
asynchronous. Unfortunately this async work is causing us some headaches
on integrating it into our API Explorer.

I started to dig in to trying to make `formdata-to-string` synchronous
and realized... why bother? We have the HAR, we can generate everything
we need for a multipart boundary from that.

* [x] Completely removes the `FormData` and `Blob` API dependencies from
this library.
* [x] Made `.convert()` a synchronous operation again.

[^1]: https://github.com/readmeio/httpsnippet/pull/199

---------

Co-authored-by: Kanad Gupta <8854718+kanadgupta@users.noreply.github.com>