fix(breaking): always return array (#225)
## 🧰 Changes

I had DX concerns about the `string[] | string` setup in the
`.convert()` method and talked offline with @erunion and agreed that it
should always return a `string[]`, so this PR makes that change. This is
a breaking change.

## 🧬 QA & Testing

Is this everything? I saw this line below but I think it's unrelated so
I left it, but let me know if I'm wrong about this.


https://github.com/readmeio/httpsnippet/blob/fc92e4000714ec677bbf712fc9537fd616e8ae65/src/targets/index.ts#L41